% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{E Function}
\usage{
E(ewq, lambda, mu, type = "")
}
\arguments{
\item{ewq}{Expected value of queue waiting time}

\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{type}{character, type in 'Obj Q' for E(Nq), 'Obj S' for E(N) or 'Wait S' for E(W)}
}
\value{
expected values
}
\description{
Depending on the type specification, the function returns the expected values of the waiting time in the system, objects in queue or objects in the system for GGC-Systems.
}
