% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWq.R
\name{EWq}
\alias{EWq}
\title{EWq Function}
\usage{
EWq(cv_lam, cv_mu, c, mu, lambda, type = "")
}
\arguments{
\item{cv_lam}{coefficient of variation of arrival rate lambda}

\item{cv_mu}{coefficient of variation of service rate mu}

\item{c}{number of server}

\item{mu}{service rate}

\item{lambda}{arrival rate}

\item{type}{character, type in Kingman for the Kingman's approximation or Whitt for the Whitt et al. approximation}
}
\value{
expected waiting time in queues
}
\description{
Returns the value of the expected value for the queue value time for GGC-Systems
!ATTENTION! cv_lam and cv_mu do NOT have to be passed squared!
}
\examples{
EWq(cv_lam=0.25, cv_mu=1, c=7, mu=1, lambda=0.5, type="Kingman")
EWq(cv_lam=0.25, cv_mu=1, c=7, mu=1, lambda=0.5, type="Whitt")
}
